AMD 5700 XT Kernel interrupts
I was noticing a lot of CPU interrupts, making the Hackintosh choppy and sluggish. Removing the WhateverGreen.kext seems to remove any of conflicting drivers used to support GPU's. Remember to remove the supporting flag in the boot section of config.plist if you own an AMD RX5700 XT GPU.